requirements.txt
Django==1.7.1
MySQL-python==1.2.4
dj-database-url==0.3.0
django-allauth==0.16.1
django-angular==0.7.10
django-compressor==1.4
django-embed-video==0.11
django-extensions==1.4.8
django-guardian==1.1.1
django-livereload==1.1
django-mathfilters==0.1.3
django-mptt==0.6.1
django-nose==1.2
django-object-actions==0.5.1
django-pdb==0.4.1
django-redis-cache==0.13.0
django-rest-auth==0.3.3
django-rest-swagger==0.2.6
django-reversion==1.8.0
django-s3-folder-storage==0.2
django-seo-js==0.2.4
django-smtp-ssl==1.0
django-storages==1.1.8
django-tagging==0.3.2
django-tastypie==0.11.1
django-userena==1.2.4
django-widget-tweaks==1.3
django-xmlrpc==0.1.5
djangorestframework==3.0.3
djangorestframework-jwt==1.1.1
-e git://github.com/pythonforfacebook/facebook-sdk.git@47b19fa23ebf943f84d81f1961dc1b9a318af539#egg=facebook_sdk-master
oauthlib==0.6.1
